requirements.txt
redis==5.0.1
beautifulsoup4==4.12.2
selenium==4.16.0
schedule==1.2.1
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
numpy==1.24.3